created a service for fetching all series
