+++ src/App.tsx
@@ -4,44 +4,19 @@ import { useEffect, useState } from 'react';
 
 import "./app.css";
 
+import { getAllSeries } from './Series/service/getAllSeries';
+
 import { ISeries, Series } from './Series';
-import AxiosSingleton from './singleton/AxiosSingleton';
 
 function App() {
 
   const [series, setSeries] = useState<ISeries[]>([]);
 
   useEffect(() => {
-    fetchData()
+    setSeries(getAllSeries);
   }, []
   )
 
-  const fetchData = async () => {
-    const axios = new AxiosSingleton;
-    await axios.getInstance().get("http://localhost:9090/api/v1/series")
-      .then(data => {
-        let remainder = 4 - (data.data.length % 4);
-        console.table(remainder, data.data.length)
-        let vector: ISeries[] = [];
-        vector.push(...data.data)
-        for (let i = 0; i < remainder; i++) {
-          vector.push(
-            {
-              title: "",
-              lastEpisode: "",
-              nextEpisode: "",
-              numberOfSeasons: "",
-              registrationNumber: "" + i
-            }
-          )
-
-        }
-        setSeries(vector)
-        console.log(vector)
-        // setSeries([...series, ...vector])
-      })
-      .catch(error => console.error(error))
-  }
 
   return (
     <div className="app">
